+++ shell/platform/common/cpp/client_wrapper/publish.gni
@@ -71,7 +71,9 @@ template("publish_client_wrapper_core") {
                              "visibility",
                            ])
     public = core_cpp_client_wrapper_includes
-    sources = core_cpp_client_wrapper_sources + [ _wrapper_readme ]
+    sources = core_cpp_client_wrapper_sources +
+              core_cpp_client_wrapper_internal_headers + [ _wrapper_readme ] +
+              temporary_shim_files
   }
 }
