+++ shell/platform/glfw/client_wrapper/testing/stub_flutter_glfw_api.h
@@ -88,6 +88,9 @@ class StubFlutterGlfwApi {
 
   // Called for FlutterDesktopShutDownEngine.
   virtual bool ShutDownEngine() { return true; }
+
+  // Called for FlutterDesktopRegistrarEnableInputBlocking.
+  virtual void RegistrarEnableInputBlocking(const char* channel) {}
 };
 
 // A test helper that owns a stub implementation, making it the test stub for
